JG-2009-10-26-0: Bugfix:      merged minor memory leaks from V61_BRANCH,                               cleanup.                  Review:      CR Issue number: Obtained from: Submitted by: Reviewed by:

+++ source/clients/common/sge_centry_qconf.c
@@ -141,8 +141,10 @@ centry_provide_modify_context(sge_gdi_ctx_class_t *ctx, lListElem **this_elem, l
                                              SP_FORM_ASCII, filename, false);
       
       if (answer_list_output(&alp)) {
-         unlink(filename);
-         FREE(filename);
+         if (filename != NULL) {
+            unlink(filename);
+            FREE(filename);
+         }
          DRETURN(false);
       }
