+++ source/clients/common/sge_hgroup_qconf.c
@@ -165,8 +165,10 @@ static bool hgroup_provide_modify_context(sge_gdi_ctx_class_t *ctx, lListElem **
                                                      SP_DEST_TMP, SP_FORM_ASCII,
                                                      filename, false);
       if (answer_list_has_error(answer_list)) {
-         unlink(filename);
-         FREE(filename);
+         if (filename != NULL) {
+            unlink(filename);
+            FREE(filename);
+         }
          DRETURN(false);
       }
